+++ src/components/User/Cart.js
@@ -2,44 +2,56 @@ import React, { Component } from 'react';
 // import axios from '../../config/axios'
 import { Link } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { getCart, deleteCart, changeToWishlist, changeToCart, counterChange } from '../../actions/cart'
+import { deleteCart, changeToWishlist, changeToCart, counterChange, getCartOnly } from '../../actions/cart'
 import {  getUsers  } from '../../actions/users'
 import cookies from 'universal-cookie'
 import swal from 'sweetalert'
 import'../../css/cart.css'
+import axios from '../../config/axios';
+import Footer from '../Footer';
 const cookie = new cookies ()
 
 class Cart extends Component{
     state={
-        cart:[],
         qty:0,
-        cartOnly:[]
+        wishlist:[]
     }
 
     onDeleteHandler = async (product_id) => {
-        var user_id = cookie.get('idLogin')
-        await this.props.deleteCart(user_id, product_id)
-        this.getCartUser()
+
+        const confirm = window.confirm('Mau hapus?')
+
+        if(confirm){
+            var user_id = cookie.get('idLogin')
+            await this.props.deleteCart(user_id, product_id)
+            this.wishList()
+            this.onlyCart()
+        }
     }
 
     onChangeToWishlist = async (product_id) => {
         var user_id = cookie.get('idLogin')
         await this.props.changeToWishlist(user_id, product_id)
-        this.getCartUser()
+        this.wishList()
+        this.onlyCart()
     }
 
     onChangeToCart = async (product_id) => {
         var user_id = cookie.get('idLogin')
         await this.props.changeToCart(user_id, product_id)
-        this.getCartUser()
+        this.wishList()
+        this.onlyCart()
     }
-
+    
+    //plus minus qty
     handleCounterChange = async (newQty, product_id) => {
         var user_id = cookie.get('idLogin')
         await this.props.counterChange(user_id, product_id, newQty)
-        this.getCartUser()
+        this.wishList()
+        this.onlyCart()
     }
     
+    //minus qty, setstate
     onMinQtyHandler = (qtyProd, product_id) => {
         if(qtyProd === 0){
             return null
@@ -50,6 +62,7 @@ class Cart extends Component{
             })
     }
 
+    //plus qty, setstate
     onPlusQtyHandler = (qtyProd, product_id, stock) => {
         if(qtyProd >= stock){
             return swal({
@@ -63,26 +76,27 @@ class Cart extends Component{
             })
     }
 
-    getCartUser = async() => {
-        var id = cookie.get('idLogin')
-        const res= await this.props.getCart(id)
-        // console.log(res.payload.data);
-       await this.setState({cart: res.payload.data})
-    //    console.log(this.state.cart);
-       
-        // this.setState({qty: res.payload.data.qty})
+    componentDidMount = () => {
+       this.wishList()
+       this.onlyCart()
     }
 
-    componentDidMount = () => {
-        this.getCartUser()
+    //gak ke state, langsung redux
+    onlyCart = async () => {
+        var user_id = cookie.get('idLogin')
+        this.props.getCartOnly(user_id)       
+    }
+
+    //setState dulu
+    wishList = async () => {
+        var user_id = cookie.get('idLogin')
+        const res = await axios.get(`/shop/wishlist/${user_id}`)
+        // console.log(res.data);
+        return this.setState({wishlist: res.data})
     }
 
     renderCart = () => {
-        const newCart = this.state.cart.filter((item) => {
-            return item.cls === 'Cart'
-        }) 
-        // console.log(newCart);
-        return newCart.map((item, i) => {
+        return this.props.cartOnly.map((item, i) => {
             return (
                 <div className='row m-2 border-info border p-4' key={i}>
                     <div className='col-3'>
@@ -99,43 +113,41 @@ class Cart extends Component{
                     </div>
                     <div className='col-4'>
                         <div className='d-flex justify-content-center'> Qty: 
-                                <button onClick={()=> {this.onMinQtyHandler(item.qty, item.id)}}
-                                 className='border-0 bg-white'><i className="fas fa-minus"></i>
-                                </button>
-                                <input type='text' className='text-center'
+                            <button onClick={()=> {this.onMinQtyHandler(item.qty, item.id)}}
+                                className='border-0 bg-white'><i className="fas fa-minus"></i>
+                            </button>
+                            <input type='text' className='text-center'
                                 onChange={()=>{return null}}
-                                value={ item.qty } style={{maxWidth: '50px'}} ></input>
-                                <button onClick={() => {this.onPlusQtyHandler(item.qty, item.id, item.stock)}}
+                                value={ item.qty } style={{maxWidth: '50px'}}>                                    
+                            </input>
+                            <button onClick={() => {this.onPlusQtyHandler(item.qty, item.id, item.stock)}}
                                 className='border-0 bg-white'><i className="fas fa-plus"></i>
-                                </button>
+                            </button>
                         </div>
                         <div className='d-flex justify-content-center mt-2'>
                             Total: Rp. {(item.price*item.qty).toLocaleString()}
                         </div>
                         <div className='d-inline-flex'>
-                            <button onClick={() => {this.onChangeToWishlist(item.id)}} className='btnMovewishlistCart'>Move to Wishlist</button>
-                            <button onClick={() => {this.onDeleteHandler(item.id)}} className='btnDeleteCart'>Delete</button>
+                            <button onClick={() => {this.onChangeToWishlist(item.id)}} 
+                                className='btnMovewishlistCart'>Move to Wishlist
+                            </button>
+                            <button onClick={() => {this.onDeleteHandler(item.id)}} 
+                                className='btnDeleteCart'>Delete
+                            </button>
                         </div>
                     </div>
                 </div>
             )
         })
-        
     }
 
-    wishList = () => {
-        
-        const newWishlist = this.state.cart.filter((item) => {
-            return item.cls === 'Wishlist'
-        })
-        // console.log(newWishlist);
-
-        return newWishlist.map((item, i) => {
+    renderWishlist = () => {
+        return this.state.wishlist.map((item, i) => {
             return (
                 <div className='row m-2 border-info border p-4' key={i}>
                     <div className='col-3'>
                         <img className="" src={`http://localhost:2010/manageproduct/list/${item.image}`} 
-                        style={{ width: '100%' }} alt="Cardcap" />
+                            style={{ width: '100%' }} alt="Cardcap"/>
                     </div>
                     <div className='col-5'>
                         <div>
@@ -146,20 +158,24 @@ class Cart extends Component{
                         </div>
                     </div>
                     <div className='col-4'>
-                        <div className='d-flex justify-content-center'> Qty: 
-                                <button onClick={()=> {this.onMinQtyHandler(item.qty, item.id)}}
+                        <div className='d-flex justify-content-center'> Qty:
+                            <button onClick={() => { this.onMinQtyHandler(item.qty, item.id) }}
                                 className='border-0 bg-white'><i className="fas fa-minus"></i>
-                                </button>
-                                <input type='text' className='text-center' 
-                                onChange={()=>{return null}}
-                                value={item.qty} style={{maxWidth: '50px'}} ></input>
-                                <button onClick={()=> {this.onPlusQtyHandler(item.qty, item.id, item.stock)}}
+                            </button>
+                            <input type='text' className='text-center'
+                                onChange={() => { return null }}
+                                value={item.qty} style={{ maxWidth: '50px' }} ></input>
+                            <button onClick={() => { this.onPlusQtyHandler(item.qty, item.id, item.stock) }}
                                 className='border-0 bg-white'><i className="fas fa-plus"></i>
-                                </button>
+                            </button>
                         </div>
                         <div className='d-inline-flex'>
-                            <button onClick={() => {this.onChangeToCart(item.id)}} className='btnMovewishlistCart'>Add to Cart</button>
-                            <button onClick={() => {this.onDeleteHandler(item.id)}} className='btnDeleteCart'>Delete</button>
+                            <button onClick={() => {this.onChangeToCart(item.id)}} 
+                                className='btnMovewishlistCart'>Add to Cart
+                            </button>
+                            <button onClick={() => {this.onDeleteHandler(item.id)}} 
+                                className='btnDeleteCart'>Delete
+                            </button>
                         </div>
                     </div>
                 </div>
@@ -169,10 +185,7 @@ class Cart extends Component{
 
     orderSummary = () => {
         
-        const newCart = this.state.cart.filter((item) => {
-            return item.cls === 'Cart'
-        })
-
+        const newCart = this.props.cartOnly
         var totalQty = 0
         for(let i =0; i<newCart.length; i++){
             totalQty += newCart[i].qty
@@ -182,7 +195,9 @@ class Cart extends Component{
         for(let i =0; i<newCart.length; i++){
             totalBuy += newCart[i].qty*newCart[i].price
         }
-        totalBuy = totalBuy.toLocaleString()
+        totalBuy = parseInt(totalBuy).toLocaleString()
+        // console.log(typeof(totalBuy));
+        
 
         return(
             <div className='border border-info p-3 m-2'>
@@ -191,42 +206,75 @@ class Cart extends Component{
                 <div className='d-flex justify-content-center'>
                     <Link to={{
                         pathname: '/checkout'
-                    }} ><button className='btn btn-danger'>Proceed to Checkout</button></Link>
+                        }} >
+                         {totalBuy !== '0' ? (
+                            <button className='btn btn-danger'>Proceed to Checkout</button>
+                         ): null}   
+                        
+                    </Link>
                 </div>
             </div>
         )
     }
 
     render(){
-        // console.log(this.state.qty);
+        // console.log(this.state.wishlist);
+        // console.log(this.props.cartOnly);
+        var a = cookie.get('cartUser')
+        console.log(a);
+        
         return(
-            <div className='container'>
-                <h4>Shopping Cart</h4>
-                <div className='row'>
-                    <div className='col-8'>
-                        <div>
-                            {this.renderCart()}
+            <div>
+                <div className='container'>
+                    <div className='row'>
+                        <div className='col-8'>
+                            {this.props.cartOnly.length > 0 ? (
+                                <div>
+                                    <h4 className='text-center'>
+                                        <span className='border-bottom border-dark'>Shopping Cart</span>
+                                    </h4>
+                                    <div>
+                                        {this.renderCart()}
+                                    </div>
+                                </div>
+                                ): (
+                                <div>
+                                    <h3 className='text-center'>Cart Anda Kosong, <Link to='/product/all'>Click me!</Link></h3>
+                                </div>
+                            )}
+                            {this.state.wishlist.length > 0 ? (
+                                <div>
+                                    <h4 className='text-center'>
+                                        <span className='border-bottom border-dark'>Wishlist</span>
+                                    </h4>
+                                    <div>
+                                        {this.renderWishlist()}
+                                    </div>
+                                </div>
+                                ): null}
                         </div>
-                        <h4>Wishlist</h4>
-                        <div>
-                            {this.wishList()}
-                        </div>
-                    </div>
-                    <div className='col-4'>
-                        <h4 className='border border-info p-3 m-2 text-center'>Order Summary</h4>
-                        <div>
-                            {this.orderSummary()}
+                        <div className='col-4'>
+                            <h4 className='border border-info p-3 m-2 text-center'>Order Summary</h4>
+                            <div>
+                                {this.orderSummary()}
+                            </div>
                         </div>
                     </div>
                 </div>
+                    <div className='fixed-bottom'>
+                        <Footer/>
+                    </div>
             </div>
+
         )
     }
 }
 
-// const mapStateToProps = (state) => {
-//     return { carts : state.cart.carts }
-//   }
+const mapStateToProps = (state) => {
+    return { 
+            cartOnly: state.cart.cartOnly 
+        }
+  }
   
-  export default connect(null, { getCart, getUsers, deleteCart, changeToCart, changeToWishlist, counterChange })(Cart)
+  export default connect(mapStateToProps, { getUsers, deleteCart, changeToCart, changeToWishlist, counterChange, getCartOnly })(Cart)
